// webui/src/lib/gen/models/ProblemInfo.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemInfoIsConvex } from './ProblemInfoIsConvex';
import type { ProblemInfoIsLinear } from './ProblemInfoIsLinear';
import type { ProblemInfoIsTwiceDifferentiable } from './ProblemInfoIsTwiceDifferentiable';
import type { ProblemInfoScenarioKeys } from './ProblemInfoScenarioKeys';
import type { VariableDomainTypeEnum } from './VariableDomainTypeEnum';
import type { ProblemInfoConstants } from './ProblemInfoConstants';
import type { ProblemInfoTensorConstants } from './ProblemInfoTensorConstants';
import type { ProblemInfoVariables } from './ProblemInfoVariables';
import type { ProblemInfoTensorVariables } from './ProblemInfoTensorVariables';
import type { ObjectiveDB } from './ObjectiveDB';
import type { ProblemInfoConstraints } from './ProblemInfoConstraints';
import type { ProblemInfoScalarizationFuncs } from './ProblemInfoScalarizationFuncs';
import type { ProblemInfoExtraFuncs } from './ProblemInfoExtraFuncs';
import type { ProblemInfoDiscreteRepresentation } from './ProblemInfoDiscreteRepresentation';
import type { ProblemInfoSimulators } from './ProblemInfoSimulators';
import type { ProblemInfoProblemMetadata } from './ProblemInfoProblemMetadata';

/**
 * Problem info request return data.
 */
export interface ProblemInfo {
	name: string;
	description: string;
	is_convex: ProblemInfoIsConvex;
	is_linear: ProblemInfoIsLinear;
	is_twice_differentiable: ProblemInfoIsTwiceDifferentiable;
	scenario_keys: ProblemInfoScenarioKeys;
	variable_domain: VariableDomainTypeEnum;
	id: number;
	user_id: number;
	constants: ProblemInfoConstants;
	tensor_constants: ProblemInfoTensorConstants;
	variables: ProblemInfoVariables;
	tensor_variables: ProblemInfoTensorVariables;
	objectives: ObjectiveDB[];
	constraints: ProblemInfoConstraints;
	scalarization_funcs: ProblemInfoScalarizationFuncs;
	extra_funcs: ProblemInfoExtraFuncs;
	discrete_representation: ProblemInfoDiscreteRepresentation;
	simulators: ProblemInfoSimulators;
	problem_metadata: ProblemInfoProblemMetadata;
}