// webui/src/lib/gen/models/problemInfo.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemInfoIsConvex } from './problemInfoIsConvex';
import type { ProblemInfoIsLinear } from './problemInfoIsLinear';
import type { ProblemInfoIsTwiceDifferentiable } from './problemInfoIsTwiceDifferentiable';
import type { ProblemInfoScenarioKeys } from './problemInfoScenarioKeys';
import type { VariableDomainTypeEnum } from './variableDomainTypeEnum';
import type { ProblemInfoConstants } from './problemInfoConstants';
import type { ProblemInfoTensorConstants } from './problemInfoTensorConstants';
import type { ProblemInfoVariables } from './problemInfoVariables';
import type { ProblemInfoTensorVariables } from './problemInfoTensorVariables';
import type { ObjectiveDB } from './objectiveDB';
import type { ProblemInfoConstraints } from './problemInfoConstraints';
import type { ProblemInfoScalarizationFuncs } from './problemInfoScalarizationFuncs';
import type { ProblemInfoExtraFuncs } from './problemInfoExtraFuncs';
import type { ProblemInfoDiscreteRepresentation } from './problemInfoDiscreteRepresentation';
import type { ProblemInfoSimulators } from './problemInfoSimulators';
import type { ProblemInfoProblemMetadata } from './problemInfoProblemMetadata';

/**
 * Problem info request return data.
 */
export interface ProblemInfo {
	name: string;
	description: string;
	is_convex: ProblemInfoIsConvex;
	is_linear: ProblemInfoIsLinear;
	is_twice_differentiable: ProblemInfoIsTwiceDifferentiable;
	scenario_keys: ProblemInfoScenarioKeys;
	variable_domain: VariableDomainTypeEnum;
	id: number;
	user_id: number;
	constants: ProblemInfoConstants;
	tensor_constants: ProblemInfoTensorConstants;
	variables: ProblemInfoVariables;
	tensor_variables: ProblemInfoTensorVariables;
	objectives: ObjectiveDB[];
	constraints: ProblemInfoConstraints;
	scalarization_funcs: ProblemInfoScalarizationFuncs;
	extra_funcs: ProblemInfoExtraFuncs;
	discrete_representation: ProblemInfoDiscreteRepresentation;
	simulators: ProblemInfoSimulators;
	problem_metadata: ProblemInfoProblemMetadata;
}